requirements.txt
absl-py==1.4.0
agora-token-builder==1.0.0
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.6.0
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.1.0
cachetools==5.3.1
cbor2==5.4.6
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
confusable-homoglyphs==3.3.1
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.9
django-appconf==1.0.5
django-axes==6.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compressor==4.3.1
django-csp==3.7
django-csp-reports==1.9.0
django-formtools==2.5.1
django-hijack==3.4.5
django-hosts==5.2
django-htmlmin==0.11.0
django-maintenance-mode==0.21.1
django-meta==2.4.2
django-payments==2.0.0
django-phonenumber-field==7.1.0
django-ratelimit==4.1.0
django-redis==5.2.0
django-registration==3.4
django-templated-mail==1.1.1
django-timezone-field==5.0
django-waffle==4.1.0
django-webpush==0.3.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fido2==1.1.1
flatbuffers==23.5.26
future==0.18.3
gast==0.4.0
geographiclib==2.0
geopy==2.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.9.0
html5lib==1.1
http-ece==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jax==0.4.12
keras==2.12.0
kombu==5.3.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
msgpack==1.0.5
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
phonenumberslite==8.13.13
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.23.3
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-fsutil==0.14.1
python3-openid==3.2.0
pytz==2023.3
pywebpush==1.9.4
pywrap==0.1.0
rcssmin==1.1.1
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
reverse-geocode==1.4.1
rjsmin==1.2.1
rsa==4.9
scipy==1.10.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
ua-parser==0.18.0
urllib3==1.26.16
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
webauthn==1.10.1
webencodings==0.5.1
Werkzeug==2.3.6
whitenoise==6.4.0
wrapt==1.14.1
zope.interface==6.0
